refactor: move some tests too
